requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
fpdf2>=2.7.4
matplotlib
scipy
google-genai
kaleido==0.2.1
tabulate
python-docx